change SLOAD to MLOAD (#17)
<!--
Borrowed from foundry.

Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.
-->

## Motivation
Reading a state variable costs more gas than reading a local variable. 
<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? In some cases there is not a problem and this
can be
thought of as being the motivation for your change.

If your PR solves a particular issue, tag that issue.
-->

## Solution
If a state variable is read multiple times, it saves gas to cache the
variable as a local variable and then read the local variable.
<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->