change bool to uint256 (#18)
<!--
Borrowed from foundry.

Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.
-->

## Motivation
Bools use 8 bit but they only need 1 bit. And it takes 20,000 gas to set
a slot from zero to non-zero with SSTORE. So changing bool to uint can
save a lot of gas.
<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? In some cases there is not a problem and this
can be
thought of as being the motivation for your change.

If your PR solves a particular issue, tag that issue.
-->

## Solution
Use uint256 instead of bool if this bool variable takes up an entire
slot.
<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->

Co-authored-by: Ryan Ghods <ryan@ryanio.com>