Fix: Only Emit BatchMetadataUpdate When Total Supply Isn't Zero (#23)
## Motivation
If `setBaseURI` is called before minting starts (i.e. when
`totalSupply() == 0`), the `BatchMetadataUpdate` event would show
`BatchMetadataUpdate(_fromTokenId = 1, _toTokenId = 0);` which can be
deemed confusing to the users.
<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? In some cases there is not a problem and this
can be
thought of as being the motivation for your change.

If your PR solves a particular issue, tag that issue.
-->

## Solution
Only emit `BatchMetadataUpdate` when total supply isn't zero. Emitting
such events when minting hasn't occurred wouldn't make sense anyway. If
no tokens have been minted, there'd be no need to refresh token metadata
amirite?
<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->