Add non-partner extension contracts (#36)
<!--
Borrowed from foundry.

Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.
-->

## Motivation

As we extend the capabilities of OpenSea's Drops program to a wider
audience, we would like to offer developers the ability to integrate
with SeaDrop without having to rely on a direct partnership with
OpenSea. Some of these developers might want a Burn or RandomOffset
extension of `ERC721SeaDrop` that does not require an `administrator`
role

<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? In some cases there is not a problem and this
can be
thought of as being the motivation for your change.

If your PR solves a particular issue, tag that issue.
-->

## Solution
Implement `ERC721SeaDropBurnable` and `ERC721SeaDropRandomOffset` as
direct extensions of `ERC721SeaDrop`. This PR also includes a minor
revision to our DropURI schema that removes the `promoImage` field,
which we do not use in our product.

<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->