Add clone factory for lightweight deployments  (#44)
<!--
Borrowed from foundry.

Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.
-->

## Motivation

Deploying SeaDrop can be very expensive; clones are much more
lightweight, but do add calldata overhead, so have not been prioritized
until public rollout.

<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? In some cases there is not a problem and this
can be
thought of as being the motivation for your change.

If your PR solves a particular issue, tag that issue.
-->

## Solution
Add a Clone factory for lightweight deploys.

<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->

- Add ERC721SeaDropCloneable
- Add ERC721SeaDropCloneFactory
- Add tests

---------

Co-authored-by: James Wenzel <jameswenzel@users.noreply.github.com>
Co-authored-by: Ryan Ghods <ryan@ryanio.com>